+++ design-system/theme/index.tsx
@@ -1,58 +1,2 @@
-import { colors } from './colors';
-
-export { palette, colors } from './colors';
-
 export { ThemeProvider, useTheme } from './themecontext';
-
-export const fontFamilies = {
-  heading:
-    "-apple-system, BlinkMacSystemFont, 'SF Pro Display','roboto',sans-serif",
-  body:
-    'Monaco,Menlo,Consolas,"Droid Sans Mono","Inconsolata","Courier New",monospace'
-};
-
-export const spacing = {
-  none: 0,
-  xsmall: 4,
-  small: 8,
-  medium: 12,
-  large: 16,
-  xlarge: 24,
-  xxlarge: 32,
-  xxxlarge: 40
-} as const;
-
-export const breakpoints = {
-  sm: 640,
-  md: 768,
-  lg: 1024,
-  xl: 1280
-} as const;
-
-export const radii = {
-  none: 0,
-  small: 3,
-  medium: 6,
-  large: 12
-} as const;
-
-export const boxShadow = {
-  1: {
-    boxShadow: `0px 6px 12px ${colors.shadow}`
-  }
-};
-
-export const fontsizes = {
-  xsmall: '0.75rem',
-  small: '0.875rem',
-  medium: '1rem',
-  large: '1.25rem',
-  xlarge: '1.5rem',
-  xxlarge: '1.875rem',
-  xxxlarge: '2.25rem',
-  xxxxlarge: '3rem'
-} as const;
-
-export type SpacingType = keyof typeof spacing;
-export type RadiiType = keyof typeof radii;
-export type FontSizeType = keyof typeof fontsizes;
+export * from './themes';
