+++ design-system/theme/themecontext.tsx
@@ -1,14 +1,12 @@
 import React, { createContext, useContext } from 'react';
 
-import { darkColors } from './colors';
+import { themeDark } from './themes';
 
 // Context lets us pass a value deep into the component tree
 // without explicitly threading it through every component.
 // Create a context for the current theme (with "light" as the default).
 export const ThemeContext = createContext({
-  theme: {
-    colors: darkColors
-  }
+  theme: themeDark
 });
 
 export const ThemeProvider = ({ children }) => {
